:recycle: Refactor Selectors
To enable full CLI operation, serialization to JSON format is requiered
from all parts of Hausse project setup. Relying on simple strings with
some specific format was not satisfactory.

Selectors' __str__ methods are removed. Instead, a method called save
is added. It returns a description of the Selector in the shape of a
python dict. This method does not need specific implementation for any
future Selector subclass.

On occasion, Selector classes names are simplfied, and Plugins
subclasses are reworked too, mainly to enable Selectors' save method
call automatically when the whole project is saved.

This commit contains also the changes needed by the revelent existent
plugins concerned by theses changes.